+++ packages/betterer/src/context/summary.ts
@@ -1,72 +1,63 @@
-import assert from 'assert';
-import { getConfig } from '../config';
-
 import { BettererRunsΩ } from './run';
 import { BettererRuns, BettererSummary } from './types';
 
 export class BettererSummaryΩ implements BettererSummary {
-  constructor(private _runs: BettererRunsΩ, private _result: string, private _expected: string | null) {}
+  public readonly unexpectedDiff: boolean;
+  public readonly shouldWrite: boolean;
 
-  public get runs(): BettererRuns {
-    return this._runs;
-  }
-
-  public get result(): string {
-    return this._result;
-  }
-
-  public get expected(): string | null {
-    return this._expected;
-  }
-
-  public get unexpectedDiff(): boolean {
-    assert(this._result);
-    const config = getConfig();
-    const hasDiff = !!this._expected && this._expected !== this._result;
-    return hasDiff && config.ci;
+  constructor(
+    public readonly runs: BettererRunsΩ,
+    public readonly result: string,
+    public readonly expected: string | null,
+    ci: boolean
+  ) {
+    const hasDiff = !!this.expected && this.expected !== this.result;
+    this.unexpectedDiff = hasDiff && ci;
+    const expectedDiff = hasDiff && !ci;
+    this.shouldWrite = !this.expected || expectedDiff;
   }
 
   public get completed(): BettererRuns {
-    return this._runs.filter((run) => run.isComplete);
+    return this.runs.filter((run) => run.isComplete);
   }
 
   public get expired(): BettererRuns {
-    return this._runs.filter((run) => run.isExpired);
+    return this.runs.filter((run) => run.isExpired);
   }
 
   public get better(): BettererRuns {
-    return this._runs.filter((run) => run.isBetter);
+    return this.runs.filter((run) => run.isBetter);
   }
 
   public get failed(): BettererRuns {
-    return this._runs.filter((run) => run.isFailed);
+    return this.runs.filter((run) => run.isFailed);
   }
 
   public get new(): BettererRuns {
-    return this._runs.filter((run) => run.isNew && run.isRan);
+    return this.runs.filter((run) => run.isNew && run.isRan);
   }
 
   public get obsolete(): BettererRuns {
-    return this._runs.filter((run) => run.isObsolete);
+    return this.runs.filter((run) => run.isObsolete);
   }
 
   public get ran(): BettererRuns {
-    return this._runs.filter((run) => run.isRan);
+    return this.runs.filter((run) => run.isRan);
   }
 
   public get same(): BettererRuns {
-    return this._runs.filter((run) => run.isSame);
+    return this.runs.filter((run) => run.isSame);
   }
 
   public get skipped(): BettererRuns {
-    return this._runs.filter((run) => run.isSkipped);
+    return this.runs.filter((run) => run.isSkipped);
   }
 
   public get updated(): BettererRuns {
-    return this._runs.filter((run) => run.isUpdated);
+    return this.runs.filter((run) => run.isUpdated);
   }
 
   public get worse(): BettererRuns {
-    return this._runs.filter((run) => run.isWorse);
+    return this.runs.filter((run) => run.isWorse);
   }
 }