+++ packages/betterer/src/runner/index.ts
@@ -1,5 +1,3 @@
-export { BettererFileManager } from './file-manager';
-export { BettererFileResolver, BettererFileResolverΩ } from './file-resolver';
 export { BettererRunnerΩ } from './runner';
-export { BettererFileGlobs, BettererFilePaths, BettererFilePatterns, BettererRunner } from './types';
+export { BettererRunner } from './types';
 export { BettererWatcherΩ } from './watcher';