+++ packages/betterer/src/utils.ts
@@ -29,10 +29,6 @@ export function getAbsolutePath(resultsPath: string, filePath: string): string {
   return normalisedPath(path.resolve(path.dirname(resultsPath), filePath));
 }
 
-export function getRelativePath(resultsPath: string, filePath: string): string {
-  return normalisedPath(path.relative(path.dirname(resultsPath), filePath));
-}
-
 type Resolve<T> = (value: T) => void;
 type Reject = (error: Error) => void;
 export type Defer<T> = {